exitValue -> double
git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@2941 19eef36c-86f0-de11-afab-0004ac968479

+++ src/com/algoTrader/service/ActionServiceImpl.java
@@ -27,7 +27,7 @@ public class ActionServiceImpl extends ActionServiceBase {
         logger.debug("retrieveTicks end (" + (System.currentTimeMillis() - startTime) + "ms execution)");
     }
 
-    protected void handleSetExitValue(int positionId, BigDecimal exitValue) throws java.lang.Exception {
+    protected void handleSetExitValue(int positionId, double exitValue) throws java.lang.Exception {
 
         long startTime = System.currentTimeMillis();
         logger.debug("setExitValue start");
